fix: 상세페이지 들어갔다 나올때 탭바가 잠시 사라지는현상
